Enhancements to project management. The OpenRubyRMK is a multi-component project, and each component should be packaged separately. In addition to this, project-wide Raketasks exist to allow coping with all subprojects at once. Adding a new subproject, e.g. a client, is as easy as adding it to the COMPONENTS array in the toplevel Rakefile.
Please also note that versioning happens now centralized in the
toplevel directory via the CENTRAL_VERSION file. The VERSION files
in the respective subdirectories are generated from this file via
the all:bump_version task and are therefore excluded from version
control; nevertheless these files must be taken into account when
building the gems for the components as they read them to
inform about their version number. An intented side effect of this
is that all components always carry the same version number.