Improved the Request DSL heavily. Eliminated the Request::DSL class. It was causing more headache that it was useful.
Requests and responses now know about their sender, that is it
is unnecessary to reach the Client instance from method to method.

Back to methods for the DSL.

Eliminated the ugly self[:sym] construct for accessing parameters.
The Request#execute methods now gets a nice hash for this.

Added optional parameters for requests.

Added automatic checking of required arguments, rendering checks
for parameter presence obsolete.

Adapted the existing requests to the updated DSL.