Ease the request DSL.
The #answer method can now be called without having to explicitely
specify tge Request and Client instances, making the whole DSL more
concise. This is achieved by wrapping the handler code in a special
object, ActionHandler instances, that carry not only the handler code,
but can also be altered to reflect the state of the currently running
request/response, which then in turn can be accessed in and by the
handler code (although the user doesn’t see this, as the answer method
implicitely uses this feature).