Make the functional testing work much better. Closes #40.
You can now instruct Karfunkel to signal its readiness by use of
the `-S <PID>' commandline parameter. This allows the functional
tests to start up *much* more quickly and reliably. Regarding the
interface stuff mentioned in #40, I suspect this to be a general
downside of asynchronous event testing; so the best you can do
here is to leave some comments in the testing code.